+++ src/components/_server/serverAuth.ts
@@ -2,9 +2,6 @@
 
 import { cookies } from "next/dist/client/components/headers";
 
-export async function setRefreshTokenToCookie(refresh_token: string) {
-  cookies().set("refreshToken", refresh_token);
-}
 export async function getRefreshTokenFromCookie() {
   return cookies().get("refreshToken")?.value;
 }
