+++ src/components/modal/CreatePostModal/index.tsx
@@ -1,13 +1,13 @@
 import { enteredGroupsAtom } from "@app/GlobalProvider";
-import AuthCheckButton from "@components/AuthCheckButton";
+import AuthCheckButton from "@components/Button/AuthCheckButton";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import clsx from "clsx";
 import { atom, useAtom, useAtomValue, useSetAtom } from "jotai";
 import Prev from "public/assets/icons/left.svg";
 import { FormEvent, useState } from "react";
-import { Button } from "src/components/Button";
+import { Button } from "@components/Button";
 import { Input } from "src/components/Input";
-import Modal from "src/components/Modal";
+import Modal from "@components/Modal";
 import TextArea from "src/components/TextArea";
 import { PostApis } from "src/lib/apis/postApis";
 import { useGlobalModal } from "../GlobalModal";
